// src/types/vendors.ts
export type Vendor = {
  id: string
  vendor_name?: string | null
  owner_name?: string | null
  inn?: number | null
  city?: string | null
  address?: string | null
  latitude?: number | null
  longitude?: number | null
  warehouse_1c?: string | null
  price_type?: string | null
  phone?: string | null
  whatsapp?: string | null
  telegram?: string | null
  ok_link?: string | null
  vk_link?: string | null
  email?: string | null
  schedule?: string | null
  juridical_phone?: string | null
  created_at?: string | null
}